Se agrega imagen perfil.png a la carpeta imagenesservicios para que no se elimine esa carpeta

+++ Nuevo/src/Logica/ControladorProveedores.java
@@ -575,8 +575,8 @@ public class ControladorProveedores implements IControladorProveedores{
         File[]imagenes = fichero.listFiles();
         
         for(int i = 0; i < imagenes.length; i++){
-            
-            imagenes[i].delete();
+            if(!imagenes[i].getName().equalsIgnoreCase("perfil.png")) 
+                imagenes[i].delete();
         }
     }
